+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 **I'm using for some apps the same icons. If this is the case with your apps or if something other went wrong, just report it [here](https://github.com/scoute-dich/Baumann_Theme/issues) or on the [xda threat](http://forum.xda-developers.com/android/themes/cm12-theme-source-t3164482).**
 
+### 4.5 (WIP)
+
+- added: icons (thanks to @thomas31415 on xda thread)
+- improved: wallpaper fragment layout
+- fixed: fab behavior
+
+
 ### 4.4
 
 - added: icons (thanks to @thomas31415 on xda thread)